split graphs into separate files
